Refactor AvatarDropdown component: Update import paths and add ProfileButton
